feat: #76 (organiza estrutura e adiciona TS (vite))
