+++ src/background/main.ts
