+++ src/content/task/getTasks.ts
@@ -1,0 +1,16 @@
+import { getTaskInfo } from "./getTaskInfo";
+
+export function getTasks(document) {
+  try {
+    const tasks = [];
+    document.querySelectorAll('.card-inner')
+      .forEach((item) => {
+        tasks.push(getTaskInfo(item))
+      })
+
+    
+    return tasks;
+  } catch (error) {
+    return [];
+  }
+}
