+++ src/content/time/parseTime.ts
@@ -1,0 +1,6 @@
+
+
+export function parseTime(time) {
+  const [hours, minutes] = time.split('.').map(Number);
+  return hours * 60 + (minutes || 0); // convert to minutes
+}
