+++ src/popup/main.ts
@@ -1,8 +1,11 @@
+// @ts-nocheck
+
 function checkNaN(hr) {
   return hr.includes('NaN') ?  'NOT FOUND' : hr;
 }
-
+  
 document.addEventListener('DOMContentLoaded', () => {
+
   chrome.tabs.query({ active: true, currentWindow: true }, (tabs) => {
     const currentUrl = tabs[0].url;
     const allowedUrls = ['taiga'];
