+++ tsconfig.json
@@ -1,15 +1,22 @@
 {
   "compilerOptions": {
-    "outDir": "./dist",
-    "target": "ES2020",
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
+    "target": "ESNext",
+    "useDefineForClassFields": true,
+    "lib": ["ES2020", "DOM", "DOM.Iterable","ESNext"],
+    "module": "ESNext",
     "skipLibCheck": true,
-    "moduleResolution": "node",
-    "allowJs": true,
-    "types": ["vite/client"]
+
+    /* Bundler mode */
+    "moduleResolution": "Bundler",
+    "allowImportingTsExtensions": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "noEmit": true,
+    "esModuleInterop": true,
+
+    /* Linting */
+    "strict": false,
   },
-  "include": ["src"]
+  "include": ["src"],
+  "references": [{ "path": "./tsconfig.node.json" }]
 }
